Configuration improvements and git settings
