+++ frontend/src/config/integrations/index.ts
@@ -1,4 +1,5 @@
 import github from './github/config';
+import git from './git/config';
 
 export interface IntegrationConfig {
   key: string; // Unique key for the integration
@@ -8,9 +9,11 @@ export interface IntegrationConfig {
   connectComponent?: Vue.Component; // Component rendered for user to connect integration
   actionComponent?: Vue.Component; // Component rendered when integration needs user action
   statusComponent?: Vue.Component; // Component rendered to show integration status
+  connectedParamsComponent?: Vue.Component; // Component rendered to show connected integration params (repositories, channels)
   dropdownComponent?: Vue.Component; // Component rendered inside dropdown for extra options
 }
 
 export const lfIntegrations: Record<string, IntegrationConfig> = {
   github,
+  git,
 };
