+++ frontend/src/shared/drawer/drawer.vue
@@ -1,64 +1,66 @@
 <template>
-  <el-drawer
-    v-model="model"
-    :custom-class="`${customClass} ${
-      hasBorder ? 'bordered' : ''
-    } ${
-      hasPadding ? '' : 'no-padding'
-    }`"
-    :show-close="false"
-    :destroy-on-close="true"
-    :close-on-click-modal="false"
-    :size="size"
-    @close="onClose"
-  >
-    <template #header="{ close, titleId, titleClass }">
-      <div class="flex grow justify-between items-start">
-        <slot name="header">
-          <div class="h-fit">
-            <div
-              v-if="preTitle"
-              class="text-gray-600 text-2xs"
-            >
-              {{ preTitle }}
-            </div>
-            <div class="flex items-center">
-              <slot name="beforeTitle" />
-              <h5
-                :id="titleId"
-                class="text-black"
-                :class="titleClass"
+  <teleport to="body">
+    <el-drawer
+      v-model="model"
+      :custom-class="`${customClass} ${
+        hasBorder ? 'bordered' : ''
+      } ${
+        hasPadding ? '' : 'no-padding'
+      }`"
+      :show-close="false"
+      :destroy-on-close="true"
+      :close-on-click-modal="false"
+      :size="size"
+      @close="onClose"
+    >
+      <template #header="{ close, titleId, titleClass }">
+        <div class="flex grow justify-between items-start">
+          <slot name="header">
+            <div class="h-fit">
+              <div
+                v-if="preTitle"
+                class="text-gray-600 text-2xs"
               >
-                {{ title }}
-              </h5>
+                {{ preTitle }}
+              </div>
+              <div class="flex items-center">
+                <slot name="beforeTitle" />
+                <h5
+                  :id="titleId"
+                  class="text-black"
+                  :class="titleClass"
+                >
+                  {{ title }}
+                </h5>
+              </div>
+              <slot name="belowTitle" />
             </div>
-            <slot name="belowTitle" />
-          </div>
-        </slot>
+          </slot>
 
-        <div class="flex gap-3 items-center">
-          <slot name="header-label" />
-          <div class="ml-3">
-            <el-button
-              class="btn btn-link btn-link--xs btn-link--primary w-8 !h-8 hover:!no-underline group"
-              @click="close"
-            >
-              <i
-                class="ri-close-line text-lg text-gray-400 group-hover:text-primary-500"
-              />
-            </el-button>
+          <div class="flex gap-3 items-center">
+            <slot name="header-label" />
+            <div class="ml-3">
+              <el-button
+                class="btn btn-link btn-link--xs btn-link--primary w-8 !h-8 hover:!no-underline group"
+                @click="close"
+              >
+                <i
+                  class="ri-close-line text-lg text-gray-400 group-hover:text-primary-500"
+                />
+              </el-button>
+            </div>
           </div>
         </div>
-      </div>
-    </template>
+      </template>
 
-    <template #default>
-      <slot name="content" />
-    </template>
-    <template v-if="showFooter" #footer>
-      <slot name="footer" />
-    </template>
-  </el-drawer>
+      <template #default>
+        <slot name="content" />
+      </template>
+      <template v-if="showFooter" #footer>
+        <slot name="footer" />
+      </template>
+    </el-drawer>
+  </teleport>
 </template>
 
 <script setup>
